Move parser tests into the right module
