+++ bindings/rust/parser.rs
@@ -355,3 +355,112 @@ impl MarkdownParser {
         self.parse_with(&mut |byte, _| &text[byte..], old_tree)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use tree_sitter::{InputEdit, Point};
+
+    use super::*;
+
+    #[test]
+    fn inline_ranges() {
+        let code = "# title\n\nInline [content].\n";
+        let mut parser = MarkdownParser::default();
+        let mut tree = parser.parse(code.as_bytes(), None).unwrap();
+
+        let section = tree.block_tree().root_node().child(0).unwrap();
+        assert_eq!(section.kind(), "section");
+        let heading = section.child(0).unwrap();
+        assert_eq!(heading.kind(), "atx_heading");
+        let paragraph = section.child(1).unwrap();
+        assert_eq!(paragraph.kind(), "paragraph");
+        let inline = paragraph.child(0).unwrap();
+        assert_eq!(inline.kind(), "inline");
+        assert_eq!(
+            tree.inline_tree(&inline)
+                .unwrap()
+                .root_node()
+                .child(0)
+                .unwrap()
+                .kind(),
+            "shortcut_link"
+        );
+
+        let code = "# Title\n\nInline [content].\n";
+        tree.edit(&InputEdit {
+            start_byte: 2,
+            old_end_byte: 3,
+            new_end_byte: 3,
+            start_position: Point { row: 0, column: 2 },
+            old_end_position: Point { row: 0, column: 3 },
+            new_end_position: Point { row: 0, column: 3 },
+        });
+        let tree = parser.parse(code.as_bytes(), Some(&tree)).unwrap();
+
+        let section = tree.block_tree().root_node().child(0).unwrap();
+        assert_eq!(section.kind(), "section");
+        let heading = section.child(0).unwrap();
+        assert_eq!(heading.kind(), "atx_heading");
+        let paragraph = section.child(1).unwrap();
+        assert_eq!(paragraph.kind(), "paragraph");
+        let inline = paragraph.child(0).unwrap();
+        assert_eq!(inline.kind(), "inline");
+        assert_eq!(
+            tree.inline_tree(&inline)
+                .unwrap()
+                .root_node()
+                .named_child(0)
+                .unwrap()
+                .kind(),
+            "shortcut_link"
+        );
+    }
+
+    #[test]
+    fn markdown_cursor() {
+        let code = "# title\n\nInline [content].\n";
+        let mut parser = MarkdownParser::default();
+        let tree = parser.parse(code.as_bytes(), None).unwrap();
+        let mut cursor = tree.walk();
+        assert_eq!(cursor.node().kind(), "document");
+        assert!(cursor.goto_first_child());
+        assert_eq!(cursor.node().kind(), "section");
+        assert!(cursor.goto_first_child());
+        assert_eq!(cursor.node().kind(), "atx_heading");
+        assert!(cursor.goto_next_sibling());
+        assert_eq!(cursor.node().kind(), "paragraph");
+        assert!(cursor.goto_first_child());
+        assert_eq!(cursor.node().kind(), "inline");
+        assert!(cursor.goto_first_child());
+        assert_eq!(cursor.node().kind(), "shortcut_link");
+        assert!(cursor.goto_parent());
+        assert!(cursor.goto_parent());
+        assert!(cursor.goto_parent());
+        assert!(cursor.goto_parent());
+        assert_eq!(cursor.node().kind(), "document");
+    }
+
+    #[test]
+    fn table() {
+        let code = "| foo |\n| --- |\n| *bar*|\n";
+        let mut parser = MarkdownParser::default();
+        let tree = parser.parse(code.as_bytes(), None).unwrap();
+        dbg!(&tree.inline_trees());
+        let mut cursor = tree.walk();
+
+        assert_eq!(cursor.node().kind(), "document");
+        assert!(cursor.goto_first_child());
+        assert_eq!(cursor.node().kind(), "section");
+        assert!(cursor.goto_first_child());
+        assert_eq!(cursor.node().kind(), "pipe_table");
+        assert!(cursor.goto_first_child());
+        assert!(cursor.goto_next_sibling());
+        assert!(cursor.goto_next_sibling());
+        assert_eq!(cursor.node().kind(), "pipe_table_row");
+        assert!(cursor.goto_first_child());
+        assert!(cursor.goto_next_sibling());
+        assert_eq!(cursor.node().kind(), "pipe_table_cell");
+        assert!(cursor.goto_first_child());
+        assert_eq!(cursor.node().kind(), "emphasis");
+    }
+}
